Moves each Opinion collection into separate file
