+++ src/fronts/opinion/variantC/components/EditorialCollection.tsx
@@ -1,0 +1,27 @@
+import React from "react";
+import { Collection as ICollection } from "../../../../api";
+import { Heading } from "./../../../../components/Heading";
+import { Multiline } from "./../../../../components/Multiline";
+import { Card as CommentCard } from "./Card";
+
+export const EditorialCollection: React.FC<{
+    collection: ICollection;
+    salt: string;
+}> = ({ collection, salt }) => {
+    const contentOne = collection.backfill[0];
+
+    // TODO
+    return (
+        <>
+            <Multiline topPadding />
+            <Heading heading={collection.displayName} />
+
+            <CommentCard
+                content={contentOne}
+                salt={salt}
+                size={"large"}
+                shouldShowImage={true}
+            />
+        </>
+    );
+};
