+++ src/fronts/opinion/variantC/components/MediaCollection.tsx
@@ -1,0 +1,22 @@
+import React from "react";
+import { Collection as ICollection } from "../../../../api";
+import { Heading } from "./../../../../components/Heading";
+import { Multiline } from "./../../../../components/Multiline";
+import { MediaCard } from "./MediaCard";
+
+export const MediaCollection: React.FC<{
+    collection: ICollection;
+    salt: string;
+}> = ({ collection, salt }) => {
+    const items = collection.backfill.map(content => (
+        <MediaCard content={content} salt={salt} />
+    ));
+
+    return (
+        <>
+            <Multiline topPadding />
+            <Heading heading={collection.displayName} />
+            {items}
+        </>
+    );
+};
